🔧 chore(.env.local.example): add missing environment variables 🔧 chore(components/carousel.tsx): add dots to the carousel 🔧 chore(prisma/schema.prisma): update database URL
Added the following missing environment variables to the .env.local.example file:
- CLOUDINARY_NAME
- CLOUDINARY_API_KEY
- CLOUDINARY_API_SECRET
- SERVER_ROOT_URL

Added dots to the carousel in the components/carousel.tsx file to improve the user experience.

Updated the database URL in the prisma/schema.prisma file to use a local SQLite database file named "dev.db". This change ensures that the application uses the correct database for development purposes.